move Smithy4s to own dir
