+++ packages/quicktype-core/src/language/Smithy4s/constants.ts
@@ -1,0 +1,77 @@
+// Use backticks for param names with symbols
+export const invalidSymbols = [
+    ":",
+    "-",
+    "+",
+    "!",
+    "@",
+    "#",
+    "%",
+    "^",
+    "&",
+    "*",
+    "(",
+    ")",
+    ">",
+    "<",
+    "/",
+    ";",
+    "'",
+    '"',
+    "{",
+    "}",
+    ":",
+    "~",
+    "`",
+    "."
+] as const;
+
+export const keywords = [
+    "abstract",
+    "case",
+    "catch",
+    "do",
+    "else",
+    "export",
+    "false",
+    "final",
+    "finally",
+    "for",
+    "forSome",
+    "if",
+    "implicit",
+    "import",
+    "new",
+    "override",
+    "package",
+    "private",
+    "protected",
+    "return",
+    "sealed",
+    "super",
+    "this",
+    "then",
+    "throw",
+    "trait",
+    "try",
+    "true",
+    "val",
+    "var",
+    "while",
+    "with",
+    "yield",
+    "Any",
+    "Boolean",
+    "Double",
+    "Float",
+    "Long",
+    "Int",
+    "Short",
+    "System",
+    "Byte",
+    "String",
+    "Array",
+    "List",
+    "Map",
+    "Enum"
+] as const;
