+++ packages/quicktype-core/src/language/Smithy4s/index.ts
@@ -1,0 +1,2 @@
+export { SmithyTargetLanguage, smithyOptions } from "./language";
+export { Smithy4sRenderer } from "./Smithy4sRenderer";
